Amanda's js-ref-type-diagnostic
I wasn't able to complete this quiz.
I struggled a lot with this lesson when it was initially taught, so I'm
not surprised I wasn't able to complete the quiz. When this info was presented
during class, it was presented as a code along. I got behind and never
was able to catch up. I had planned to try to redo the lab on my own
over the weekend.

When this info was presented in class. The var names during the code
along were different than what was in the readme. so I was able to read the
material to catch up either.